requirements.txt
python-telegram-bot
python-telegram-bot[job-queue]
requests
dotenv
websockets
apache-airflow-providers-postgres
yfinance
pandas
matplotlib
numpy
sqlalchemy
psycopg2-binary
requests
